Revert "Geometry"
