+++ groalea/graphio.py
@@ -12,32 +12,22 @@
 # 2.5. Compute properties when it is possible (sphere, ...)
 # 2.6. 2D draw of the graph
 
-
-# 3. Add enum like FUNCTIONAL
-
 from StringIO import StringIO
 from math import radians
 from math import sqrt
 from math import cos
 from math import sin
-from copy import deepcopy
+
 import xml.etree.ElementTree as xml
 
 from openalea.core.graph.property_graph import PropertyGraph
 import openalea.plantgl.all as pgl
-
-from .geometry import (TurtleState, FunctionalGeometry, rgb_color,
-                       is_matrix, transform4, frame,
-                       orientation, project3Dto2D, determinant, no_interior,
-                       grotation, directionalTropism, orthogonalTropism, adjust_lu)
-
 Vector3 = pgl.Vector3
 Vector4 = pgl.Vector4
 Color4Array = pgl.Color4Array
 
 
 class RootedGraph(PropertyGraph):
-    """ A general graph with a root vertex. """
 
     def _set_root(self, root):
         self._root = root
@@ -46,8 +36,6 @@ class RootedGraph(PropertyGraph):
         return self._root
 
     root = property(_get_root, _set_root)
-
-
 
 
 class Parser(object):
@@ -59,21 +47,20 @@ class Parser(object):
                   'F', 'F0', 'M', 'M0', 'RL', 'RU', 'RH', 'V', 'RV', 'RV0', 'RG', 'RD', 'RO', 'RP', 'RN', 'AdjustLU',
                   'L', 'LAdd', 'LMul', 'D', 'DAdd', 'DMul', 'P', 'Translate', 'Scale', 'Rotate']
 
-    FUNCTIONAL = -10
-
     def parse(self, fn):
         self.trash = []
         self._graph = None
         self._scene = None
 
         # Turtle intialisation
-        # self._turtle_diameter = -1.
-        # self._turtle_color = []
-        # self._turtle_color_setflag = False
-        # self._turtle_length = -1.
-        # self._turtle_tropism = 0.
-        # self._tropism_drectionList = []
-        # self._tropism_target = []
+        self._turtle_diameter = -1.
+        self._turtle_color = []
+        self._turtle_color_setflag = False
+        self._turtle_length = -1.
+        self._turtle_tropism = 0.
+        self._tropism_drectionList = []
+        self._tropism_target = []
+        self._translateX = self._translateY = self._translateZ = 0.
 
         # print "pass parse intialisation"
         doc = xml.parse(fn)
@@ -81,10 +68,8 @@ class Parser(object):
 
         self.has_type = False
         self.types(doc.findall('type'))
-
         # self.guiding_edgelist_produce(doc.findall('edge'))
         # print "#####pass to dispatch(root)"
-
         self.dispatch(root)
         self.scenegraph()
 
@@ -103,11 +88,11 @@ class Parser(object):
             list(elt)
         # print 'Dispatch get **', elt.getchildren()
         # return self.__getattribute__(elt.tag)(elt.getchildren(), **elt.attrib)
-        #try:
-        return self.__getattribute__(elt.tag)(list(elt), **elt.attrib)
-#        except Exception, e:
-#            print e
-#            raise Exception("Invalid element %s" % elt.tag)
+        try:
+            return self.__getattribute__(elt.tag)(list(elt), **elt.attrib)
+        except Exception, e:
+            print e
+            raise Exception("Invalid element %s" % elt.tag)
 
     def dispatch2(self, method_name, args):
         try:
@@ -121,23 +106,18 @@ class Parser(object):
         A graph is a set of nodes and edges.
         """
         print "pass graph(self, elements) function"
-
         graph = self._graph = RootedGraph()
-
         self._edges = {}
         graph._types = self._types
 
-        # Add initial properties
         graph.add_vertex_property("name")
         graph.add_vertex_property("type")
         graph.add_vertex_property("parameters")
         graph.add_vertex_property("color")
         graph.add_vertex_property("geometry")
         graph.add_vertex_property("transform")
-        graph.add_vertex_property("turtle_state")
         graph.add_edge_property("edge_type")
 
-        # Process / Parse all the xml nodes contains directly inside the <graph> node.
         for elt in elements:
             self.dispatch(elt)
 
@@ -195,15 +175,7 @@ class Parser(object):
         self._graph.root = int(root_id)
 
     def node(self, properties, id, type, name=None):
-        """
-        TODO: Write an exhaustive list of examples here
 
-        <node id="3" name="" type="L">
-            <property name="length" value="6.0"/>
-        </node>
-
-
-        """
         # print "node id, type, name = ", id, type, name
         # print "current turtle_length = ", self._turtle_length
 
@@ -212,8 +184,6 @@ class Parser(object):
             name = str(id)
 
         graph = self._graph
-
-        self._current_turtle = TurtleState()
 
         graph.add_vertex(id)
         if name:
@@ -229,12 +199,9 @@ class Parser(object):
         properties2 = [p for p in properties
                        if p.attrib['name'] not in ('transform', 'color')]
 
-        # TODO : Improve the design
-
+        args = self._get_args(properties2)
         if type in ['P', 'PointCloud']:
             args = self._get_args(properties)
-        else:
-            args = self._get_args(properties2)
 
         graph.vertex_property('parameters')[id] = args
 
@@ -244,24 +211,18 @@ class Parser(object):
         else:
             shape, transfo = self.dispatch2(type, args)
 
-        # End of TODO
-
         assert len(transfos) <= 1
-
         if transfos:
             transfo = self.transform(transfos[0].getchildren())
 
         color = None
-
-        # Store the turtle state for this node, ONLY
         if colors:
             if len(colors) > 0:
                 if len(colors[0].getchildren()) != 0:
                     color = self.color(colors[0].getchildren())
 
-                self._current_turtle.color = color
-                # if self._turtle_color_setflag is True:
-                #    color = self._turtle_color
+                if self._turtle_color_setflag is True:
+                    color = self._turtle_color
             else:
                 raise Exception("color is null!!!")
 
@@ -271,11 +232,6 @@ class Parser(object):
             graph.vertex_property('transform')[id] = transfo
         if color:
             graph.vertex_property('color')[id] = color
-
-        if self._current_turtle:
-            graph.vertex_property('turtle_state')[id] = self._current_turtle
-
-        self._current_turtle = None
 
     Node = node
 
@@ -343,17 +299,16 @@ class Parser(object):
         return (pgl.PointSet(v3array, c4array, pointSize), None)
 
     def Polygon(self, vertices, **kwds):
-        """ TODO: Move code to geometry """
         points = str(vertices)
         points = [float(num) for num in points.split(",")]
         items, chunk = points, 3
         p3list = zip(*[iter(items)] * chunk)
-        p2list = project3Dto2D(p3list)
+        p2list = self.project3Dto2D(p3list)
         pd2list = []
         for i in range(len(p2list)):
             pd2list.append({i: p2list[i]})
         indexlist = []
-        poly_orientation = orientation(p2list)
+        poly_orientation = self.orientation(p2list)
 
         while len(pd2list) >= 3:
             for cur in range(len(pd2list)):
@@ -363,8 +318,8 @@ class Parser(object):
                 # we will iterate at end only if poly_orientation
                 # was incorrect.
                 pcur, pprev, pnex = pd2list[cur].values()[0], pd2list[prev].values()[0], pd2list[nex].values()[0]
-                det = determinant(pcur, pprev, pnex)
-                inside = no_interior(pprev, pcur, pnex, pd2list, poly_orientation)
+                det = self.determinant(pcur, pprev, pnex)
+                inside = self.no_interior(pprev, pcur, pnex, pd2list, poly_orientation)
                 if (det == poly_orientation) and inside:
                     # Same orientation as polygon
                     # No points inside
@@ -375,6 +330,90 @@ class Parser(object):
                     del(pd2list[cur])
                     break
         return (pgl.TriangleSet(pgl.Point3Array(p3list), indexlist), None)
+
+    def project3Dto2D(self, p3list):
+        v01 = Vector3((p3list[1][0] - p3list[0][0]), (p3list[1][1] - p3list[0][1]), (p3list[1][2] - p3list[0][2]))
+        v12 = Vector3((p3list[2][0] - p3list[1][0]), (p3list[2][1] - p3list[1][1]), (p3list[2][2] - p3list[1][2]))
+        vn = pgl.cross(v01, v12)
+
+        p2s = []
+        # cosTheta = A dot B/(|A|*|B|) => if A dot B ==0, then Theta == 90
+        # if polygon not || y axis, project it to the y=0 plane
+        if pgl.dot(vn, Vector3(0, 1, 0)) != 0:
+            for i in range(len(p3list)):
+                v = p3list[i][0], p3list[i][2]
+                p2s.append(v)
+
+        else:
+            # if polygon || y axis and z axis (it will perpendicular x axis), project it to the x=0 plane
+            # if polygon || y axis and x axis (it will perpendicular z axis), project it to the z=0 plane
+            # if polygon || y axis, not || x and z (it will not perpendicular z and x
+            # axis), project it to the z=0 plane (or x=0 plane)
+            if pgl.dot(vn, Vector3(0, 0, 1)) == 0:
+                for i in range(len(p3list)):
+                    v = p3list[i][1], p3list[i][2]
+                    p2s.append(v)
+
+            else:
+                for i in range(len(p3list)):
+                    v = p3list[i][0], p3list[i][1]
+                    p2s.append(v)
+        return p2s
+
+    def orientation(self, v):
+        area = 0.0
+        # Compute the area (times 2) of the polygon
+        for i in range(len(v)):
+            area += v[i - 1][0] * v[i][1] - v[i - 1][1] * v[i][0]
+
+        if area >= 0.0:
+            return 0
+        return 1
+
+    @staticmethod
+    def determinant(self, p1, p2, p3):
+        """ TODO : Use static method here
+        """
+        determ = (p2[0] - p1[0]) * (p3[1] - p1[1]) - (p3[0] - p1[0]) * (p2[1] - p1[1])
+
+        if determ >= 0:
+            return 1
+        return 0
+
+    @staticmethod
+    def no_interior(p1, p2, p3, v, poly_or):
+        """ TODO : Use static method
+        """
+        for p in v:
+            if p.values()[0] == p1 or p.values()[0] == p2 or p.values()[0] == p3:
+                # Don't bother checking against yourself
+                continue
+
+            if self.determinant(p1, p2, p.values()[0]) == poly_or or \
+                    self.determinant(p3, p1, p.values()[0]) == poly_or or \
+                    self.determinant(p2, p3, p.values()[0]) == poly_or:
+                # This point is outside
+                continue
+            # The point is inside
+            return False
+        # No points inside this triangle
+        return True
+
+    @staticmethod
+    def IsSameSide(A, B, C, P):
+        """ TODO: Use CamelCase."""
+        AB = B - A
+        AC = C - A
+        AP = P - A
+        v1 = pgl.cross(AB, AC)
+        v2 = pgl.cross(AB, AP)
+        # v1 and v2 should point to the same direction
+        return pgl.dot(v1, v2) >= 0
+
+    @staticmethod
+    def PointinTriangle(A, B, C, P):
+        """ TODO: Use CamelCase."""
+        return IsSameSide(A, B, C, P) and IsSameSide(B, C, A, P) and IsSameSide(C, A, B, P)
 
     sphere = Sphere
     box = Box
@@ -388,51 +427,55 @@ class Parser(object):
     # Turtle implementation:
     # F0, M, M0, RV, RG, AdjustLU
     def F(self, length=1., diameter=-1., color=14, **kwds):
-
-        def f(turtle):
-            height = turtle.length
-            radius = turtle.diameter /2.
-            color = turtle.color
-            return (pgl.Cylinder(radius=radius, height=height),
-            pgl.Matrix4.translation(Vector3(0, 0, height)))
-
-        length = float(length)
+        height = float(length)
+        self._turtle_length = height
         diameter = float(diameter)
-        color3 = rgb_color(color)
-
-        self._current_turtle.length = length
-        self._current_turtle.diameter = diameter
-        self._current_turtle.set_diameter = True
-        self._current_turtle.color = color3
-
-        return (FunctionalGeometry(f), self.FUNCTIONAL)
+        if diameter == -1.:
+            diameter = self._turtle_diameter
+            if diameter == -1:
+                diameter = self._turtle_diameter = 0.1
+        radius = diameter / 2.
+        ega_color = [(0, 0, 0),
+                     (0, 0, 170),
+                     (0, 170, 0),
+                     (0, 170, 170),
+                     (170, 0, 0),
+                     (170, 0, 170),
+                     (170, 85, 0),
+                     (170, 170, 170),
+                     (85, 85, 85),
+                     (85, 85, 255),
+                     (85, 255, 85),
+                     (85, 255, 255),
+                     (255, 85, 85),
+                     (255, 85, 255),
+                     (255, 255, 85),
+                     (255, 255, 255), ]
+        color3 = pgl.Color3(ega_color[int(color)])
+        self._turtle_color = color3
+        self._turtle_color_setflag = True
+        return (pgl.Cylinder(radius=radius, height=height),
+                pgl.Matrix4.translation(Vector3(0, 0, height)))
 
     def F0(self, **kwds):
-        def f(turtle):
-            height = turtle.length
-            radius = turtle.diameter /2.
-            color = turtle.color
-            return (pgl.Cylinder(radius=radius, height=height),
-                    pgl.Matrix4.translation(Vector3(0, 0, height)))
-
-        self._current_turtle.set_diameter = True
-        self._current_turtle.set_length = True
-
-        return (FunctionalGeometry(f), self.FUNCTIONAL)
+        height = self._turtle_length
+        if self._turtle_diameter == -1:
+            self._turtle_diameter = 0.1
+        radius = self._turtle_diameter / 2.
+        if height <= 0:
+            height = 1.
+        return (pgl.Cylinder(radius=radius, height=height),
+                pgl.Matrix4.translation(Vector3(0, 0, height)))
 
     def M(self, length=1., **kwds):
         height = float(length)
         return (None, pgl.Matrix4.translation(Vector3(0, 0, height)))
 
     def M0(self, **kwds):
-
-        def f(turtle):
-            height = turtle.length
-            return (None,
-                    pgl.Matrix4.translation(Vector3(0, 0, height)))
-
-        self._current_turtle.set_length = True
-        return (FunctionalGeometry(f), self.FUNCTIONAL)
+        height = self._turtle_length
+        if height <= 0:
+            height = 1.
+        return (None, pgl.Matrix4.translation(Vector3(0, 0, height)))
 
     def RL(self, angle, **kwds):
         # Rotation around the x axis
@@ -453,12 +496,12 @@ class Parser(object):
         return (None, pgl.Matrix4(matrix))
 
     def V(self, argument=0., **kwds):
-        self._current_turtle.tropism = float(argument)
+        self._turtle_tropism = float(argument)
         return (None, None)
 
     def RV(self, argument=1., **kwds):
         """ Gravitropism. """
-        self._current_turtle.tropism = float(argument)
+        self._turtle_tropism = float(argument)
         return (None, -2)
 
     def RV0(self, **kwds):
@@ -466,26 +509,32 @@ class Parser(object):
 
     def RG(self, **kwds):
         """ Maximal gravitropism such that local z-direction points downwards. """
-        self._current_turtle.tropism = 1e10
+        self._turtle_tropism = 1e10
         return (None, -2)
 
     def RD(self, direction, strength, **kwds):
-        self._current_turtle.tropism = float(strength)
+        self._turtle_tropism = float(strength)
         direction = str(direction)
-        self._current_turtle.direction = tuple([float(num) for num in direction.split(",")])
+        self._tropism_drectionList = [float(num) for num in direction.split(",")]
         return (None, -3)
 
     def RO(self, direction, strength, **kwds):
-        self.RD(direction, strength, **kwds)
+        self._turtle_tropism = float(strength)
+        direction = str(direction)
+        self._tropism_drectionList = [float(num) for num in direction.split(",")]
         return (None, -4)
 
     def RP(self, target, strength, **kwds):
-        self._current_turtle.tropism = float(strength)
+        self._turtle_tropism = float(strength)
         target = str(target)
-        self._current_turtle.tropism_target = tuple([float(num) for num in target.split(",")])
+        self._tropism_target = [float(num) for num in target.split(",")]
         return (None, -5)
 
-    RN = RP
+    def RN(self, target, strength, **kwds):
+        self._turtle_tropism = float(strength)
+        t = str(target)
+        self._tropism_target = [float(num) for num in t.split(",")]
+        return (None, -5)
 
     def AdjustLU(self, **kwds):
         """ Rotate around local z-axis such that local y-axis points upwards as far as possible."""
@@ -493,46 +542,62 @@ class Parser(object):
 
     def L(self, length=1., **kwds):
         """ Set the turtle state to the given length. """
-        self._current_turtle.length = float(length)
+        self._turtle_length = float(length)
         return (None, None)
 
     def LAdd(self, argument=1., **kwds):
-        self._current_turtle.length_add = float(argument)
+        self._turtle_length += float(argument)
         return (None, None)
 
     def LMul(self, argument=1., **kwds):
-        self._current_turtle.length_mul = float(argument)
+        self._turtle_length *= float(argument)
         return (None, None)
 
     def D(self, diameter=0.1, **kwds):
         """ Set the turtle state to the given diameter. """
-        self._current_turtle.diameter = float(diameter)
+        self._turtle_diameter = float(diameter)
         return (None, None)
 
     def DAdd(self, argument=1., **kwds):
-        self._current_turtle.diameter_add = float(argument)
+        self._turtle_diameter += float(argument)
         return (None, None)
 
     def DMul(self, argument=1., **kwds):
-        self._current_turtle.diameter_mul = float(argument)
+        self._turtle_diameter *= float(argument)
         return (None, None)
 
     def P(self, color=14, **kwds):
         """ Set the turtle state to the given color. """
-        color3 = rgb_color(color)
-        self._current_turtle.color = color3
+        ega_color = [(0, 0, 0),
+                     (0, 0, 170),
+                     (0, 170, 0),
+                     (0, 170, 170),
+                     (170, 0, 0),
+                     (170, 0, 170),
+                     (170, 85, 0),
+                     (170, 170, 170),
+                     (85, 85, 85),
+                     (85, 85, 255),
+                     (85, 255, 85),
+                     (85, 255, 255),
+                     (255, 85, 85),
+                     (255, 85, 255),
+                     (255, 255, 85),
+                     (255, 255, 255), ]
+        color3 = pgl.Color3(ega_color[int(color)])
+        self._turtle_color = color3
+        self._turtle_color_setflag = True
+        #Parser.turtle_color = color3
+        #Parser.turtle_color_setflag = True
         return (None, None)
 
     def Translate(self, translateX=0., translateY=0., translateZ=0., **kwds):
-        # TODO: put the code in geometry
         tx = float(translateX)
         ty = float(translateY)
         tz = float(translateZ)
         return (None, pgl.Matrix4.translation(Vector3(tx, ty, tz)))
 
     def Scale(self, scaleX=0., scaleY=0., scaleZ=0., **kwds):
-        # TODO: put the code in geometry
-
         sx = float(scaleX)
         sy = float(scaleY)
         sz = float(scaleZ)
@@ -540,8 +605,6 @@ class Parser(object):
         return (None, pgl.Matrix4(matrix))
 
     def Rotate(self, rotateX=0., rotateY=0., rotateZ=0., **kwds):
-        # TODO: put the code in geometry
-
         rx = radians(float(rotateX))
         ry = radians(float(rotateY))
         rz = radians(float(rotateZ))
@@ -551,9 +614,12 @@ class Parser(object):
         matrix = mx * my * mz
         return (None, pgl.Matrix4(matrix))
 
-    def transform(self, elements, **kwds):
-        # TODO: put the code in geometry
+    def turtle_color(self, color):
+        color = int(color)
+        if 0 <= color <= 15:
+            color = 14
 
+    def transform(self, elements, **kwds):
         matrix = elements[0]
         assert matrix.tag == 'matrix'
         m4 = map(float, matrix.text.strip().split())
@@ -565,7 +631,6 @@ class Parser(object):
         rgb = elements[0]
         assert rgb.tag == 'rgb'
         color = pgl.Color3(*(int(float(x) * 255) for x in rgb.text.strip().split()))
-
         return color
 
     def edge(self, elements, src_id, dest_id, type, id=None):
@@ -636,12 +701,8 @@ class Parser(object):
 
         edge_type = g.edge_property("edge_type")
         transform = g.vertex_property("transform")
-        local_turtles = g.vertex_property("turtle_state")
 
         transfos = {g.root: pgl.Matrix4()}
-
-        # CPL
-        turtles = {g.root: TurtleState()}
 
         def parent(vid):
             for eid in g.in_edges(vid):
@@ -649,16 +710,167 @@ class Parser(object):
                     return g.source(eid)
             return vid
 
+        def grotation(m, strength):
+            t = m.getColumn(2)
+            v0 = t.x
+            v1 = t.y
+            v2 = t.z
+            q = 1 / sqrt(t.x * t.x + t.y * t.y + t.z * t.z)
+            v0 *= q
+            v1 *= q
+            v2 *= q
+            v02 = v0 * v0
+            v12 = v1 * v1
+            q = v02 + v12
+            m00 = m11 = m22 = m10 = m20 = m01 = m21 = m02 = m12 = 0
+            local_m = 0
+            if (q < 1e-10) or (v2 * v2 > 0.99999):
+                if v2 * (v2 - self._turtle_tropism) < 0:
+                    m00 = m.getRow(0).x
+                    m10 = -m.getRow(1).x
+                    m20 = -m.getRow(2).x
+                    m01 = m.getRow(0).y
+                    m11 = -m.getRow(1).y
+                    m21 = -m.getRow(2).y
+                    m02 = m.getRow(0).z
+                    m12 = -m.getRow(1).z
+                    m22 = -m.getRow(2).z
+                else:
+                    m00 = m11 = m22 = 1
+                    m10 = m20 = m01 = m21 = m02 = m12 = 0
+            else:
+                n = 1 / sqrt(1 - 2 * strength * v2 + strength * strength)
+                m22 = (1 - strength * v2) * n
+                m02 = strength * v0 * n
+                m20 = -m02
+                m12 = strength * v1 * n
+                m21 = -m12
 
-        def update_turtle(v, ts):
-            local_turtle = local_turtles.get(v, TurtleState())
-            global_turtle = ts.combine(local_turtle)
-            return global_turtle
+                q = 1 / q
+                m00 = (v12 + v02 * m22) * q
+                m11 = (v02 + v12 * m22) * q
+                m01 = m10 = v0 * v1 * (m22 - 1) * q
+
+            vec1 = Vector4(m00, m10, m20, 0)
+            vec2 = Vector4(m01, m11, m21, 0)
+            vec3 = Vector4(m02, m12, m22, 0)
+            vec4 = Vector4(0, 0, 0, 1)
+            local_m = pgl.Matrix4(vec1, vec2, vec3, vec4)
+            return local_m
+
+        def invTransformVector(t, v):
+            x = v[0]
+            y = v[1]
+            m00 = t.getRow(0).x
+            m01 = t.getRow(0).y
+            m02 = t.getRow(0).z
+            m10 = t.getRow(1).x
+            m11 = t.getRow(1).y
+            m12 = t.getRow(1).z
+            m20 = t.getRow(2).x
+            m21 = t.getRow(2).y
+            m22 = t.getRow(2).z
+            d0 = m11 * m22 - m12 * m21
+            d1 = m12 * m20 - m10 * m22
+            d2 = m10 * m21 - m11 * m20
+            d = 1.0 / (m00 * d0 + m01 * d1 + m02 * d2)
+            v[0] = d0 * d * x + (m21 * m02 - m01 * m22) * d * y + (m01 * m12 - m02 * m11) * d * v[2]
+            v[1] = d1 * d * x + (m00 * m22 - m02 * m20) * d * y + (m10 * m02 - m00 * m12) * d * v[2]
+            v[2] = d2 * d * x + (m20 * m01 - m00 * m21) * d * y + (m00 * m11 - m01 * m10) * d * v[2]
+            return
+
+        def setFromAxisAngle(x, y, z, angle):
+            n = sqrt(x * x + y * y + z * z)
+            n = 1 / n
+            x *= n
+            y *= n
+            z *= n
+            c = cos(angle)
+            s = sin(angle)
+            omc = 1.0 - c
+            m00 = c + x * x * omc
+            m11 = c + y * y * omc
+            m22 = c + z * z * omc
+            tmp1 = x * y * omc
+            tmp2 = z * s
+            m01 = tmp1 - tmp2
+            m10 = tmp1 + tmp2
+            tmp1 = x * z * omc
+            tmp2 = y * s
+            m02 = tmp1 + tmp2
+            m20 = tmp1 - tmp2
+            tmp1 = y * z * omc
+            tmp2 = x * s
+            m12 = tmp1 - tmp2
+            m21 = tmp1 + tmp2
+            m03 = m13 = m23 = m30 = m31 = m32 = 0
+            m33 = 1
+            vec1 = Vector4(m00, m10, m20, m30)
+            vec2 = Vector4(m01, m11, m21, m31)
+            vec3 = Vector4(m02, m12, m22, m32)
+            vec4 = Vector4(m03, m13, m23, m33)
+            return pgl.Matrix4(vec1, vec2, vec3, vec4)
+
+        def directionalTropism(m, direction, strength):
+            t = m.getColumn(2)
+            x = direction[2] * t.y - direction[1] * t.z
+            y = direction[0] * t.z - direction[2] * t.x
+            z = direction[1] * t.x - direction[0] * t.y
+            vec3 = Vector3(x, y, z)
+            angle = strength * sqrt((x * x + y * y + z * z) / (t.x * t.x + t.y * t.y + t.z * t.z))
+            if (angle * angle) >= 1e-20:
+                invTransformVector(m, vec3)
+                return setFromAxisAngle(vec3.x, vec3.y, vec3.z, angle)
+            else:
+                vec1 = Vector4(1, 0, 0, 0)
+                vec2 = Vector4(0, 1, 0, 0)
+                vec3 = Vector4(0, 0, 1, 0)
+                vec4 = Vector4(0, 0, 0, 1)
+                return pgl.Matrix4(vec1, vec2, vec3, vec4)
+
+        def orthogonalTropism(m, direction, strength):
+            t = m.getColumn(2)
+            x = direction[2] * t.y - direction[1] * t.z
+            y = direction[0] * t.z - direction[2] * t.x
+            z = direction[1] * t.x - direction[0] * t.y
+            vec3 = Vector3(x, y, z)
+            angle = -strength * (t.x * direction[0] + t.y * direction[1] + t.z *
+                                 direction[2]) / sqrt(t.x * t.x + t.y * t.y + t.z * t.z)
+
+            if (angle * angle) >= 1e-20:
+                invTransformVector(m, vec3)
+                return setFromAxisAngle(vec3.x, vec3.y, vec3.z, angle)
+            else:
+                vec1 = Vector4(1, 0, 0, 0)
+                vec2 = Vector4(0, 1, 0, 0)
+                vec3 = Vector4(0, 0, 1, 0)
+                vec4 = Vector4(0, 0, 0, 1)
+                return pgl.Matrix4(vec1, vec2, vec3, vec4)
+
+        def positionalTropism(m, target, strength):
+            x = target[0] - m.getRow(0).w
+            y = target[1] - m.getRow(1).w
+            z = target[2] - m.getRow(2).w
+            l = x * x + y * y + z * z
+            t = m.getColumn(2)
+            if l > 0:
+                xv = z * t.y - y * t.z
+                yv = x * t.z - z * t.x
+                zv = y * t.x - x * t.y
+                vec3 = Vector3(xv, yv, zv)
+                angle = strength * sqrt((xv * xv + yv * yv + zv * zv) / (l * (t.x * t.x + t.y * t.y + t.z * t.z)))
+                if (angle * angle) >= 1e-20:
+                    invTransformVector(m, vec3)
+                    return setFromAxisAngle(vec3.x, vec3.y, vec3.z, angle)
+            else:
+                vec1 = Vector4(1, 0, 0, 0)
+                vec2 = Vector4(0, 1, 0, 0)
+                vec3 = Vector4(0, 0, 1, 0)
+                vec4 = Vector4(0, 0, 0, 1)
+                return pgl.Matrix4(vec1, vec2, vec3, vec4)
 
         for v in breadth_first_search(g, vid):
-            pid = parent(v)
-
-            if pid == v and v != g.root:
+            if parent(v) == v and v != g.root:
                 print "ERRRORRRR"
                 print v
                 continue
@@ -666,44 +878,51 @@ class Parser(object):
             # print "parent(v)", parent(v)
             # print "transfos", transfos
             #m = transfos[parent(v)]
-            m = transfos.get(pid)
-
-            # CPL
-            ts = turtles.get(pid, TurtleState())
-            gt =global_turtle = update_turtle(v, ts)
-
-            local_t = transform.get(v)
-            if local_t == self.FUNCTIONAL:
-                # Get the functional shape to compute the transformation and geometry
-                # with the turtle state
-                local_t = self.f_shape(v, global_turtle)
-
+            m = transfos.get(parent(v))
 
             # print "every m : ", m
             # Transform the current shape with the stack of transfos m from the root.
             # Store the result in the graph.
-            self._local2global(v, m, global_turtle.color)
-
-
+            self._local2global(v, m)
+            local_t = transform.get(v)
 
             # print "local_t : ", local_t
             if local_t == -1:
-                m = adjust_lu(m)
+                # TODO : AdjustLU
+                # debug
+                #frame(m, self._scene, color=1)
+                t = m.getColumn(3)
+                t = Vector3(t.x, t.y, t.z)
+
+                m3 = pgl.Matrix3(m)
+                x, y, z = Vector3(1, 0, 0), Vector3(0, 1, 0), Vector3(0, 0, 1)
+                X, Y, Z = m * Vector4(1, 0, 0, 0), m * Vector4(0, 1, 0, 0), m * Vector4(0, 0, 1, 0)
+                Z = Vector3(Z.x, Z.y, Z.z)
+                new_x = z ^ Z
+                if pgl.normSquared(new_x) > 1e-3:
+                    new_y = Z ^ new_x
+                    new_x.normalize()
+                    new_y.normalize()
+                    m = pgl.BaseOrientation(new_x, new_y).getMatrix()
+                    m = m.translation(t) * m
+                    #frame(m, self._scene, color=2)
+                else:
+                    print 'AdjustLU: The two vectors are Colinear'
             elif local_t == -2:
                 #RV and RG
-                local_m = grotation(m, gt.tropism)
+                local_m = grotation(m, self._turtle_tropism)
                 m = m * local_m
             elif local_t == -3:
                 # RD
-                local_m = directionalTropism(m, gt.tropism_direction, gt.tropism)
+                local_m = directionalTropism(m, self._tropism_drectionList, self._turtle_tropism)
                 m = m * local_m
             elif local_t == -4:
                 # RO
-                local_m = orthogonalTropism(m, gt.tropism_direction, gt.tropism)
+                local_m = orthogonalTropism(m, self._tropism_drectionList, self._turtle_tropism)
                 m = m * local_m
             elif local_t == -5:
                 #RP and RN
-                local_m = positionalTropism(m, gt.tropism_target, gt.tropism)
+                local_m = positionalTropism(m, self._tropism_target, self._turtle_tropism)
                 m = m * local_m
             elif local_t:
                 if local_t.getColumn(3) != Vector4(0, 0, 0, 1):
@@ -713,9 +932,10 @@ class Parser(object):
             else:
                 # print m
                 pass
-
             transfos[v] = m
-            turtles[v] = global_turtle
+
+            if parent(v) == 7295 and v == 181:
+                print m.getColumn(3)
 
     def traverse(self, vid, transfos):
         if vid in self.visited:
@@ -744,24 +964,13 @@ class Parser(object):
                 for new_vid in self.traverse(target_vid, [m]):
                     yield new_vid
 
-    def f_shape(self, vid, t):
-        g = self._graph
-        geometry = g.vertex_property("geometry")
-        transform = g.vertex_property("transform")
-
-        shape = geometry.get(vid)
-        geom, transfo = shape(t)
-
-        geometry[vid] = geom
-        transform[vid] = transfo
-        return transfo
-
-    def _local2global(self, vid, matrix, color):
+    def _local2global(self, vid, matrix):
         g = self._graph
         geometry = g.vertex_property("geometry")
         colors = g.vertex_property("color")
         final_geometry = g.vertex_property("final_geometry")
         shape = geometry.get(vid)
+        color = colors.get(vid)
         edge_type = g.edge_property("edge_type")
 
         if shape:
@@ -772,11 +981,55 @@ class Parser(object):
             shape.id = vid
             final_geometry[vid] = shape
 
-        if color:
-            colors[vid] = color
-
     def _get_args(self, properties):
         return dict([(p.attrib['name'], p.attrib['value']) for p in properties])
+
+
+def is_matrix(shape):
+    return type(shape) == pgl.Matrix4
+
+
+def transform4(matrix, shape):
+    """
+    Return a shape transformed by a Matrix4.
+    """
+    scale, (a, e, r), translation = matrix.getTransformation2()
+    shape = pgl.Translated(translation,
+                           pgl.Scaled(scale,
+                                      pgl.EulerRotated(a, e, r,
+                                                       shape)))
+    return shape
+
+
+##########################################################################
+# Debug utility
+
+def frame(matrix, scene, color=1):
+    """ Add a frame to the scene.
+    The frame is represented by the matrix.
+    :param color: allow to distinguish between to identical frames.
+    """
+    if color == 1:
+        r = pgl.Material(pgl.Color3(*(255, 0, 0)))
+        g = pgl.Material(pgl.Color3(*(0, 255, 0)))
+        b = pgl.Material(pgl.Color3(*(0, 0, 255)))
+    else:
+        r = pgl.Material(pgl.Color3(*(255, 0, 255)))
+        g = pgl.Material(pgl.Color3(*(255, 255, 0)))
+        b = pgl.Material(pgl.Color3(*(0, 0, 0)))
+
+    cylinder = pgl.Cylinder(radius=0.005, height=1)
+    #x = pgl.AxisRotated(Vector3(0,1,0), radians(90), cylinder)
+    #y = pgl.AxisRotated(Vector3(1,0,0), radians(-90), cylinder)
+    z = cylinder
+
+    #geom_x = transform4(matrix, x)
+    #scene.add(pgl.Shape(geom_x, r))
+    #geom_y = transform4(matrix, y)
+    #scene.add(pgl.Shape(geom_y, g))
+    geom_z = transform4(matrix, z)
+    scene.add(pgl.Shape(geom_z, b))
+##########################################################################
 
 
 class Dumper(object):
