--- share/data/DMul_Branch/DMul_Branch_3Run_Interface_XEG.xml
@@ -1,108 +1,0 @@
-<graph>
-  <root root_id="1"/>
-  <node id="13" name="" type="F0">
-    <property name="color">
-      <rgb>1.0 1.0 0.33333334</rgb>
-    </property>
-  </node>
-  <edge id="12" src_id="11" dest_id="13" type="successor"/>
-  <node id="3" name="" type="L">
-    <property name="length" value="1.0"/>
-  </node>
-  <edge id="2" src_id="1" dest_id="3" type="successor"/>
-  <node id="14" name="" type="RU">
-    <property name="angle" value="30.0"/>
-  </node>
-  <edge id="13" src_id="13" dest_id="14" type="branch"/>
-  <node id="15" name="" type="DMul">
-    <property name="argument" value="0.5"/>
-  </node>
-  <edge id="14" src_id="14" dest_id="15" type="successor"/>
-  <node id="16" name="" type="RU">
-    <property name="angle" value="-30.0"/>
-  </node>
-  <edge id="15" src_id="25" dest_id="16" type="branch"/>
-  <node id="4" name="" type="D">
-    <property name="diameter" value="0.2"/>
-  </node>
-  <edge id="3" src_id="3" dest_id="4" type="successor"/>
-  <node id="10" name="" type="F0">
-    <property name="color">
-      <rgb>1.0 1.0 0.33333334</rgb>
-    </property>
-  </node>
-  <edge id="9" src_id="4" dest_id="10" type="successor"/>
-  <node id="11" name="" type="DMul">
-    <property name="argument" value="0.5"/>
-  </node>
-  <edge id="10" src_id="6" dest_id="11" type="successor"/>
-  <type name="A">
-    <extends name="Sphere"/>
-  </type>
-  <node id="17" name="" type="A">
-    <property name="radius" value="0.3"/>
-    <property name="color">
-      <rgb>0.0 1.0 0.0</rgb>
-    </property>
-  </node>
-  <edge id="16" src_id="22" dest_id="17" type="successor"/>
-  <node id="18" name="" type="RU">
-    <property name="angle" value="30.0"/>
-  </node>
-  <edge id="17" src_id="25" dest_id="18" type="branch"/>
-  <node id="12" name="" type="DMul">
-    <property name="argument" value="0.5"/>
-  </node>
-  <edge id="11" src_id="7" dest_id="12" type="successor"/>
-  <node id="19" name="" type="DMul">
-    <property name="argument" value="0.5"/>
-  </node>
-  <edge id="18" src_id="18" dest_id="19" type="successor"/>
-  <node id="20" name="" type="A">
-    <property name="radius" value="0.3"/>
-    <property name="color">
-      <rgb>0.0 1.0 0.0</rgb>
-    </property>
-  </node>
-  <edge id="19" src_id="15" dest_id="20" type="successor"/>
-  <node id="21" name="" type="RU">
-    <property name="angle" value="-30.0"/>
-  </node>
-  <edge id="20" src_id="13" dest_id="21" type="branch"/>
-  <node id="22" name="" type="DMul">
-    <property name="argument" value="0.5"/>
-  </node>
-  <edge id="21" src_id="21" dest_id="22" type="successor"/>
-  <node id="23" name="" type="DMul">
-    <property name="argument" value="0.5"/>
-  </node>
-  <edge id="22" src_id="16" dest_id="23" type="successor"/>
-  <node id="24" name="" type="A">
-    <property name="radius" value="0.3"/>
-    <property name="color">
-      <rgb>0.0 1.0 0.0</rgb>
-    </property>
-  </node>
-  <edge id="23" src_id="19" dest_id="24" type="successor"/>
-  <node id="6" name="" type="RU">
-    <property name="angle" value="30.0"/>
-  </node>
-  <edge id="5" src_id="10" dest_id="6" type="branch"/>
-  <node id="7" name="" type="RU">
-    <property name="angle" value="-30.0"/>
-  </node>
-  <edge id="6" src_id="10" dest_id="7" type="branch"/>
-  <node id="25" name="" type="F0">
-    <property name="color">
-      <rgb>1.0 1.0 0.33333334</rgb>
-    </property>
-  </node>
-  <edge id="24" src_id="12" dest_id="25" type="successor"/>
-  <node id="26" name="" type="A">
-    <property name="radius" value="0.3"/>
-    <property name="color">
-      <rgb>0.0 1.0 0.0</rgb>
-    </property>
-  </node>
-  <edge id="25" src_id="23" dest_id="26" type="successor"/>
-</graph>
