--- share/data/DMul_Branch/DMul_Branch_Test_XL.txt
@@ -1,10 +1,0 @@
-module A extends Sphere(0.3)
-{
-	{setShader(GREEN);}
-}
-
-public void run ()
-[
-	Axiom ==> L(1)D(0.2) A;
-	A ==> F0 [RU(30) DMul(0.5) A] [RU(-30) DMul(0.5) A ];
-]
