--- share/data/DMul_Branch/readme.txt
@@ -1,5 +1,0 @@
-The file DMul_Branch_Test_XL.txt is a XL rule for this test case, which contain DMul commands within branches.
-The file DMul_Branch_3Run_GroIMP.jpg is the 3D result after running the XL rule 3 times on GroIMP platform.  
-The file DMul_Branch_3Run_Interface_XEG.xml is a XEG generated by interfacing package on GroIMP side, by which a wrong 3D result (DMul_Branch_3Run_Interface.jpg) will be pruduced.
-
-Note: DMul_Branch_3Run_Interface_XEG.xml is just one possible resulting XEG generated by interfacing package on GroIMP side. Although these XEGs have different sequential order in the sense of XML, they indeed indicate the same tree graph sturcture (Guide by edge's properties: src_id, dest_id). However the interfacing package on OpenAlea side produces different 3D results depends on the XML sequential order, not the tree graph sturcture.
