--- share/data/LMul/LMul_2Run_Interface_XEG.xml
@@ -1,42 +1,0 @@
-<graph>
-  <root root_id="1"/>
-  <node id="6" name="" type="RU">
-    <property name="angle" value="90.0"/>
-  </node>
-  <edge id="5" src_id="8" dest_id="6" type="branch"/>
-  <node id="7" name="" type="RH">
-    <property name="angle" value="180.0"/>
-  </node>
-  <edge id="6" src_id="9" dest_id="7" type="successor"/>
-  <node id="3" name="" type="F0">
-    <property name="color">
-      <rgb>1.0 1.0 0.33333334</rgb>
-    </property>
-  </node>
-  <edge id="2" src_id="4" dest_id="3" type="successor"/>
-  <node id="4" name="" type="L">
-    <property name="length" value="5.0"/>
-  </node>
-  <edge id="3" src_id="1" dest_id="4" type="successor"/>
-  <node id="8" name="" type="LMul">
-    <property name="argument" value="0.5"/>
-  </node>
-  <edge id="7" src_id="3" dest_id="8" type="successor"/>
-  <node id="9" name="" type="F0">
-    <property name="color">
-      <rgb>1.0 1.0 0.33333334</rgb>
-    </property>
-  </node>
-  <edge id="8" src_id="8" dest_id="9" type="successor"/>
-  <type name="A">
-    <extends name="Node"/>
-  </type>
-  <node id="10" name="" type="A"/>
-  <edge id="9" src_id="7" dest_id="10" type="successor"/>
-  <node id="11" name="" type="F0">
-    <property name="color">
-      <rgb>1.0 1.0 0.33333334</rgb>
-    </property>
-  </node>
-  <edge id="10" src_id="6" dest_id="11" type="successor"/>
-</graph>
