--- share/data/LMul/Readme.txt
@@ -1,7 +1,0 @@
-The file LMul_Test_XL.txt is a XL rule for this test case.
-The file LMul_2Run_GroIMP.jpg is the 3D result after running the XL rule 2 times on GroIMP platform.  
-The file LMul_2Run_Interface_XEG.xml is a XEG generated by interfacing package on GroIMP side, by which a wrong 3D result (LMul_2Run_Interface.jpg) will be pruduced.
-The file LMul_Branch_5Run_GroIMP.jpg is the 3D result after running the XL rule 5 times on GroIMP platform. 
-The file LMul_Branch_5Run_Interface.jpg is a wrong 3D result after running the XL rule 5 times on interface.   
-
-Note: LMul_2Run_Interface_XEG.xml is just one possible resulting XEG generated by interfacing package on GroIMP side after 2 times running of the rule. Although these XEGs have different sequential order in the sense of XML, they indeed indicate the same tree graph sturcture (Guide by edge's properties: src_id, dest_id). However the interfacing package on OpenAlea side produces different 3D results depends on the XML sequential order, not the tree graph sturcture.
