--- share/data/LMul_Branch/LMul_Branch_3Run_Interface_XEG.xml
@@ -1,104 +1,0 @@
-<graph>
-  <root root_id="1"/>
-  <node id="12" name="" type="LMul">
-    <property name="argument" value="0.8"/>
-  </node>
-  <edge id="11" src_id="23" dest_id="12" type="successor"/>
-  <node id="13" name="" type="RU">
-    <property name="angle" value="30.0"/>
-  </node>
-  <edge id="12" src_id="14" dest_id="13" type="branch"/>
-  <node id="7" name="" type="LMul">
-    <property name="argument" value="0.8"/>
-  </node>
-  <edge id="6" src_id="11" dest_id="7" type="successor"/>
-  <node id="14" name="" type="F0">
-    <property name="color">
-      <rgb>1.0 1.0 0.33333334</rgb>
-    </property>
-  </node>
-  <edge id="13" src_id="7" dest_id="14" type="successor"/>
-  <node id="15" name="" type="LMul">
-    <property name="argument" value="0.8"/>
-  </node>
-  <edge id="14" src_id="13" dest_id="15" type="successor"/>
-  <node id="3" name="" type="L">
-    <property name="length" value="1.0"/>
-  </node>
-  <edge id="2" src_id="1" dest_id="3" type="successor"/>
-  <node id="16" name="" type="LMul">
-    <property name="argument" value="0.8"/>
-  </node>
-  <edge id="15" src_id="25" dest_id="16" type="successor"/>
-  <type name="A">
-    <extends name="Sphere"/>
-  </type>
-  <node id="17" name="" type="A">
-    <property name="radius" value="0.1"/>
-    <property name="color">
-      <rgb>0.0 1.0 0.0</rgb>
-    </property>
-  </node>
-  <edge id="16" src_id="15" dest_id="17" type="successor"/>
-  <node id="18" name="" type="A">
-    <property name="radius" value="0.1"/>
-    <property name="color">
-      <rgb>0.0 1.0 0.0</rgb>
-    </property>
-  </node>
-  <edge id="17" src_id="19" dest_id="18" type="successor"/>
-  <node id="19" name="" type="LMul">
-    <property name="argument" value="0.8"/>
-  </node>
-  <edge id="18" src_id="20" dest_id="19" type="successor"/>
-  <node id="20" name="" type="RU">
-    <property name="angle" value="-30.0"/>
-  </node>
-  <edge id="19" src_id="14" dest_id="20" type="branch"/>
-  <node id="8" name="" type="RU">
-    <property name="angle" value="-30.0"/>
-  </node>
-  <edge id="7" src_id="10" dest_id="8" type="branch"/>
-  <node id="21" name="" type="F0">
-    <property name="color">
-      <rgb>1.0 1.0 0.33333334</rgb>
-    </property>
-  </node>
-  <edge id="20" src_id="9" dest_id="21" type="successor"/>
-  <node id="10" name="" type="F0">
-    <property name="color">
-      <rgb>1.0 1.0 0.33333334</rgb>
-    </property>
-  </node>
-  <edge id="9" src_id="3" dest_id="10" type="successor"/>
-  <node id="22" name="" type="A">
-    <property name="radius" value="0.1"/>
-    <property name="color">
-      <rgb>0.0 1.0 0.0</rgb>
-    </property>
-  </node>
-  <edge id="21" src_id="16" dest_id="22" type="successor"/>
-  <node id="23" name="" type="RU">
-    <property name="angle" value="-30.0"/>
-  </node>
-  <edge id="22" src_id="21" dest_id="23" type="branch"/>
-  <node id="24" name="" type="A">
-    <property name="radius" value="0.1"/>
-    <property name="color">
-      <rgb>0.0 1.0 0.0</rgb>
-    </property>
-  </node>
-  <edge id="23" src_id="12" dest_id="24" type="successor"/>
-  <node id="25" name="" type="RU">
-    <property name="angle" value="30.0"/>
-  </node>
-  <edge id="24" src_id="21" dest_id="25" type="branch"/>
-  <node id="11" name="" type="RU">
-    <property name="angle" value="30.0"/>
-  </node>
-  <edge id="10" src_id="10" dest_id="11" type="branch"/>
-  <node id="9" name="" type="LMul">
-    <property name="argument" value="0.8"/>
-  </node>
-  <edge id="8" src_id="8" dest_id="9" type="successor"/>
-</graph>
