--- share/data/LMul_Branch/LMul_Branch_Test_XL.txt
@@ -1,10 +1,0 @@
-module A extends Sphere(0.1)
-{
-	{setShader(GREEN);}
-}
-
-public void run ()
-[
-	Axiom ==> L(1) A;
-	A ==> F0 [RU(30) LMul(0.8) A] [RU(-30) LMul(0.8) A ];
-]
