--- share/data/P/P_Test_XL.txt
@@ -1,11 +1,0 @@
-module A extends Sphere(0.2)
-{
-	{setShader(GREEN);}
-}
-
-
-public void run ()
-[
-	Axiom ==> A;
-	A ==> P(14) F(1) [RU(30) A] RU(-30) A;
-]
