--- share/data/P/Readme.txt
@@ -1,7 +1,0 @@
-The file P_Test_XL.txt is a XL rule for this test case.
-The file P_3Run_GroIMP.jpg is the 3D result after running the XL rule 3 times on GroIMP platform.  
-The file P_3Run_Interface_XEG.xml is a XEG generated by interfacing package on GroIMP side, by which a wrong 3D result (P_3Run_Interface.jpg) will be pruduced.
-The file P_4Run_GroIMP.jpg is the 3D result after running the XL rule 4 times on GroIMP platform. 
-The file P_4Run_Interface.jpg is a wrong 3D result after running the XL rule 4 times on interface.   
-
-Note: P_3Run_Interface_XEG.xml is just one possible resulting XEG generated by interfacing package on GroIMP side after 3 times running of the rule. Although these XEGs have different sequential order in the sense of XML, they indeed indicate the same tree graph sturcture (Guide by edge's properties: src_id, dest_id). However the interfacing package on OpenAlea side produces different 3D results depends on the XML sequential order, not the tree graph sturcture.
