--- share/data/length/L_XEG_w.xml
@@ -1,13 +1,0 @@
-<graph>
-  <root root_id="1"/>
-  <node id="3" name="" type="F0">
-    <property name="color">
-      <rgb>1.0 1.0 0.33333334</rgb>
-    </property>
-  </node>
-  <edge id="2" src_id="4" dest_id="3" type="successor"/>
-  <node id="4" name="" type="L">
-    <property name="length" value="6.0"/>
-  </node>
-  <edge id="3" src_id="1" dest_id="4" type="successor"/>
-</graph>
