--- share/data/length/readme.txt
@@ -1,6 +1,0 @@
-The file L_Test_XL.txt is the XL rule for this test case.
-The file L_Run_GroIMP.jpg is the 3D result after running the XL rule on GroIMP platform.
-The file L_XEG_w.xml is a XEG generated by interfacing package on GroIMP side, by which a wrong 3D result (L_Run_Interface_w.jpg) will be pruduced. 
-The file L_XEG_r.xml is a XEG generated by interfacing package on GroIMP side, by which a right 3D result (L_Run_Interface_r.jpg) will be pruduced. 
-
-Note: both L_XEG_w.xml and L_XEG_r.xml are generated by interfacing package on GroIMP side, while their xml node order is not sequentially the same, they indicate the same tree graph sturcture (Guide by edge's properties: src_id, dest_id).
